+++ lib/nostr_mail.dart
@@ -10,6 +10,8 @@ export 'src/constants.dart';
 export 'src/models/attachment_ref.dart';
 export 'src/models/email.dart' show Email;
 export 'src/models/encrypted_blob.dart';
+export 'src/models/gift_wrap_state.dart'
+    show FailedGiftWrap, GiftWrapFailure, GiftWrapProgress, GiftWrapStage;
 export 'src/models/mail_event.dart';
 export 'src/models/private_settings.dart' show PrivateSettings;
 export 'src/models/recipient.dart'